Move AutoDisposePlugins to lifecycle artifact, as they're only used here
