+++ lifecycle/autodispose-lifecycle/src/main/java/com/uber/autodispose/lifecycle/AutoDisposePlugins.java
@@ -1,11 +1,11 @@
 /*
- * Copyright (c) 2017. Uber Technologies
+ * Copyright (C) 2018. Uber Technologies
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.uber.autodispose;
+package com.uber.autodispose.lifecycle;
 
 import io.reactivex.functions.Consumer;
 import org.jetbrains.annotations.Nullable;
