+++ lifecycle/autodispose-lifecycle/src/test/java/com/uber/autodispose/lifecycle/LifecycleScopeProviderCompletableTest.java
@@ -16,7 +16,6 @@
 
 package com.uber.autodispose.lifecycle;
 
-import com.uber.autodispose.AutoDisposePlugins;
 import com.uber.autodispose.RxErrorsRule;
 import com.uber.autodispose.test.RecordingObserver;
 import io.reactivex.Completable;
