+++ lifecycle/autodispose-lifecycle/src/test/java/com/uber/autodispose/lifecycle/LifecycleScopeProviderParallelFlowableTest.java
@@ -17,7 +17,6 @@
 package com.uber.autodispose.lifecycle;
 
 import com.uber.autodispose.AutoDispose;
-import com.uber.autodispose.AutoDisposePlugins;
 import com.uber.autodispose.RxErrorsRule;
 import io.reactivex.Flowable;
 import io.reactivex.processors.PublishProcessor;
